Update Android.mk
Over thought it. Everything is now correct.

Don't forget to add vendor_available: true to the 2 bp files.
One bp file is libnbaio. Scroll down untill you that modeule and the it. The other bp file is libsonic in the sonic folder. The path will be found in the external folder